requirements.txt
audio-recorder-streamlit
streamlit
openai-whisper
google-generativeai